+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 All notable changes to `csotto` will be documented in this file.
 
+## v1.2.0 - 2024-10-25
+- Utilizes Otto 41.1.3
+- Adds flag -m, which incorporates a newly introduced simplified function to download objects in-memory, instead of downloading them blockwise
+- Adjustments in READMEs and help text
+
+**Full Changelog**: https://github.com/rechtlogisch/csotto/commits/v1.2.0
+
 ## v1.1.0 - 2024-07-29
 - Utilizes Otto 40.2.8
 - Supports most security tokens (Abholzertifikat not implemented in this demo)
 - Default certificate now: test-softorg-pse.pfx
+
+**Full Changelog**: https://github.com/rechtlogisch/csotto/commits/v1.1.0
 
 ## v1.0.0 - 2024-06-08
 - Initial version